+++ interview/models.py
@@ -209,7 +209,12 @@ class Offer(models.Model):
 
 class ProcessManager(models.Manager):
     def for_user(self, user):
-        q = super().get_queryset().filter(start_date__gte=user.date_joined)
+        q = (
+            super()
+            .get_queryset()
+            .filter(Q(start_date__gte=user.date_joined) | Q(responsible__in=[user.consultant]))
+            .distinct()
+        )
         if user.consultant.is_external:
             q = q.filter(sources=user.consultant.limited_to_source)
         return q
@@ -441,7 +446,12 @@ class InterviewKind(models.Model):
 
 class InterviewManager(models.Manager):
     def for_user(self, user):
-        q = super(InterviewManager, self).get_queryset().filter(process__start_date__gte=user.date_joined)
+        q = (
+            super(InterviewManager, self)
+            .get_queryset()
+            .filter(Q(process__start_date__gte=user.date_joined) | Q(interviewers__in=[user.consultant]))
+            .distinct()
+        )
         if user.consultant.is_external:
             q = q.filter(process__sources=user.consultant.limited_to_source)
         return q