+++ interview/tests.py
@@ -88,7 +88,7 @@ class AccessRestrictionDateTestCase(TestCase):
         self.p.start_date = datetime.datetime(2016, 10, 10, tzinfo=datetime.timezone.utc)
         self.p.save()
         self.i = InterviewFactory(process=self.p)
-        self.i.interviewers.set([self.consultantOld, self.consultantNew])
+        self.i.interviewers.set([self.consultantOld])
         # self.i.save()
 
     def test_view_process(self):
@@ -1261,3 +1261,42 @@ class ImportCognitoFormTestCase(TestCase):
 
         documents = Document.objects.filter(candidate=candidate)
         self.assertEqual(documents.count(), 0)
+
+
+class SeeLinkedProcessCreatedBeforeUserJoinedTestCase(TestCase):
+    def setUp(self):
+        self.subsidiary = SubsidiaryFactory()
+        self.consultant = ConsultantFactory(company=self.subsidiary)
+
+        self.client.force_login(self.consultant.user)
+
+    def test_general_behaviour(self):
+        # create some processes and assert that they are correctly displayed
+        processes = []
+        for i in range(5):
+            tmp = ProcessFactory()
+            tmp.responsible.add(self.consultant)
+            tmp.save()
+            processes.append(tmp)
+
+        response = self.client.get(reverse("process-list"))
+        self.assertEqual(response.status_code, 200)
+
+        open_processes = response.context["open_processes_table"].data
+        self.assertEqual(len(open_processes), 5)
+        for process in processes:
+            self.assertTrue(process in open_processes)
+
+    def test_linked_process_created_before_user_joined_is_displayed(self):
+        older_process = ProcessFactory()
+        # create a process older that consultant
+        older_process.start_date = self.consultant.user.date_joined - relativedelta(months=1)
+        older_process.responsible.add(self.consultant)
+
+        response = self.client.get(reverse("process-list"))
+        self.assertEqual(response.status_code, 200)
+
+        # assert process is displayed in list even if it's outisde consultant's scope of view
+        open_processes = response.context["open_processes_table"].data
+        self.assertEqual(len(open_processes), 1)
+        self.assertTrue(older_process in open_processes)