+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Add global subsidiary filter
 - Add goal to new interview mail notification
 - Fix a bug where non-unique subsidiary codes could be created by its factory
+- Fix a bug where an interviewer was unable to see linked processes older than him
 
 # v1.16.0 (2022-11-14)
 - More refined factories